Pull upstream

+++ truffle/mx.truffle/mx_truffle.py
@@ -67,11 +67,12 @@ def javadoc(args, vm=None):
     mx.javadoc(['--unified'] + args)
     javadocDir = os.sep.join([_suite.dir, 'javadoc'])
     index = os.sep.join([javadocDir, 'index.html'])
-    indexContent = open(index, 'r').read()
-    indexContent = indexContent.replace('src="allclasses-frame.html"', 'src="com/oracle/truffle/api/vm/package-frame.html"')
-    indexContent = indexContent.replace('src="overview-summary.html"', 'src="com/oracle/truffle/tutorial/package-summary.html"')
-    new_file = open(index, "w")
-    new_file.write(indexContent)
+    if exists(index):
+        indexContent = open(index, 'r').read()
+        indexContent = indexContent.replace('src="allclasses-frame.html"', 'src="com/oracle/truffle/api/vm/package-frame.html"')
+        indexContent = indexContent.replace('src="overview-summary.html"', 'src="com/oracle/truffle/tutorial/package-summary.html"')
+        new_file = open(index, "w")
+        new_file.write(indexContent)
     checkLinks(javadocDir)
 
 def checkLinks(javadocDir):
